implement linear data generation for 2d space, also fix linspace

+++ src/utilities.cpp
@@ -1,13 +1,15 @@
 #include "utilities.h"
 
 #include <cassert>
+#include <ctime>
+#include <sys/time.h>
 
 std::vector<float> linspace(float start, float stop, unsigned int num)
 {
     assert(num > 0);
 
-    std::vector<double> ret(num);
-    double dx = num > 1 ? (stop-start)/(num-1) : 0;
+    std::vector<float> ret(num);
+    float dx = num > 1 ? (stop-start)/(num-1) : 0;
     for (int i = 0; i < num; ++i) ret[i] = start + i*dx;
 
     return ret;
